+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - MariaDB 11.7+: Support vector columns and indexes
 - PostgreSQL 9.2+: Highlight queries in process list
 - SQLite: Preserve STRICT and WITHOUT ROWID in alter table
+- SQLite: Do not treat INTEGER PRIMARY KEY without AUTOINCREMENT as auto increment
 - MS SQL: Fix multiple result sets (bug #1180)
 - New plugin: Redis driver
 - New plugin: Login behind a reverse proxy